Move interface files to separate directory.
