+++ Generation/Generation/IHepMCMergeTool.h
